requirements.txt
# ============================================================================
# WatchVine WhatsApp Bot - Python Dependencies
# ============================================================================

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# MongoDB
pymongo==4.6.1

# Google AI
google-generativeai==0.8.3

# HTTP Requests
requests==2.31.0

# Web Scraping
beautifulsoup4==4.12.2
lxml==5.1.0

# Environment Variables
python-dotenv==1.0.0

# Google Sheets API (Optional)
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
gspread==5.12.3
oauth2client==4.1.3

# Data Processing
numpy==1.26.3

# Image Processing
Pillow==10.2.0

# Utilities
python-dateutil==2.8.2

# Async Support
aiohttp==3.9.1

# Image Search (CLIP + FAISS)
sentence-transformers==2.5.1
huggingface-hub==0.20.3
faiss-cpu==1.7.4
imagehash==4.3.1
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
transformers==4.37.2
torch==2.1.2
torchvision==0.16.2
